Fix CI: isolate enrollment popover matcher from tRPC module
Tests imported enrollmentHistory.ts which eagerly loads the tRPC client.
Move findDefaultEnrollmentHistoryIndex and context types to
enrollmentHistoryPopoverMatch.ts so Vitest does not pull trpc/fetch.
